+++ presentation/src/main/java/by/grodno/vasili/presentation/dagger/AppComponent.java
@@ -10,26 +10,13 @@ import java.util.Map;
 import javax.inject.Provider;
 import javax.inject.Singleton;
 
-import by.grodno.vasili.data.datasource.FirebaseNoteEntityDatasource;
-import by.grodno.vasili.data.datasource.NoteEntityDatasource;
-import by.grodno.vasili.data.entity.mapper.NoteEntityDataMapper;
-import by.grodno.vasili.data.repository.NoteDataRepository;
-import by.grodno.vasili.domain.interactor.GetNotesListUseCase;
-import by.grodno.vasili.domain.repository.NoteRepository;
 import by.grodno.vasili.presentation.NoteApplication;
-import by.grodno.vasili.presentation.dagger.scopes.NotesActivityScope;
-import by.grodno.vasili.presentation.feature.notes.NotesActivity;
-import by.grodno.vasili.presentation.feature.notes.NotesViewModel;
-import by.grodno.vasili.presentation.thread.IOThread;
-import by.grodno.vasili.presentation.thread.UIThread;
 import dagger.Binds;
 import dagger.BindsInstance;
 import dagger.Component;
 import dagger.Module;
 import dagger.Provides;
-import dagger.android.ContributesAndroidInjector;
 import dagger.android.support.AndroidSupportInjectionModule;
-import dagger.multibindings.IntoMap;
 
 @Singleton
 @Component(modules = {AndroidSupportInjectionModule.class, AppModule.class, ActivityBuilder.class})
@@ -58,31 +45,3 @@ interface AppModule {
     }
 }
 
-@Module
-interface ActivityBuilder {
-    @SuppressWarnings("unused")
-    @NotesActivityScope
-    @ContributesAndroidInjector(modules = {NotesActivityModule.class})
-    NotesActivity bindNotesActivity();
-}
-
-@Module
-abstract class NotesActivityModule {
-
-    @Provides
-    @NotesActivityScope
-    static GetNotesListUseCase provideUserCase() {
-        NoteEntityDataMapper mapper = new NoteEntityDataMapper();
-        NoteEntityDatasource datasource = new FirebaseNoteEntityDatasource(mapper);
-        NoteRepository repository = new NoteDataRepository(datasource, mapper);
-        UIThread uiThread = new UIThread();
-        IOThread ioThread = new IOThread();
-        return new GetNotesListUseCase(ioThread, uiThread, repository);
-    }
-
-    @SuppressWarnings("unused")
-    @Binds
-    @IntoMap
-    @ViewModelKey(NotesViewModel.class)
-    abstract ViewModel bindNotesViewModel(NotesViewModel notesViewModel);
-}
