+++ presentation/src/main/java/by/grodno/vasili/presentation/dagger/ViewModelKey.java
@@ -12,6 +12,6 @@ import dagger.MapKey;
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @MapKey
-@interface ViewModelKey {
+public @interface ViewModelKey {
     Class<? extends ViewModel> value();
 }
