--- presentation/src/main/java/by/grodno/vasili/presentation/feature/notes/NoteItem.java
@@ -1,16 +1,0 @@
-package by.grodno.vasili.presentation.feature.notes;
-
-/**
- * Model for present Note
- */
-class NoteItem {
-    String id;
-    String title;
-    String description;
-
-    NoteItem(String id, String title, String description) {
-        this.id = id;
-        this.title = title;
-        this.description = description;
-    }
-}
