+++ presentation/src/main/java/by/grodno/vasili/presentation/model/NoteItem.java
@@ -1,0 +1,16 @@
+package by.grodno.vasili.presentation.model;
+
+/**
+ * Model for present Note
+ */
+public class NoteItem {
+    public String id;
+    public String title;
+    public String description;
+
+    public NoteItem(String id, String title, String description) {
+        this.id = id;
+        this.title = title;
+        this.description = description;
+    }
+}
